Simplify / cleanup
